requirements.txt
# 严格锁定版本
flask==2.2.5
werkzeug==2.2.3  # 必须精确匹配
flask-sqlalchemy==3.0.3
pyjwt==2.4.0
gmssl==3.2.2
python-dotenv==0.19.0
gunicorn==20.1.0
pymysql==1.1.1
Flask-WTF==0.15.1
bleach==4.1.0
flask_mail==0.9.1
itsdangerous==2.1.0
